+++ ftmcloud/api/domains/attributes/controllers/controller.py
@@ -6,7 +6,7 @@ from ftmcloud.api.domains.attributes.services.attribute_services import Attribut
 
 from ftmcloud.models.patchdocument import PatchDocument
 from ftmcloud.models.response import Response, ResponseWithHttpInfo
-from ftmcloud.models.attribute import Attribute
+from ftmcloud.models.domains.attributes.attribute import Attribute
 from ftmcloud.utils.views import controller
 
 router = APIRouter()
