+++ ftmcloud/api/domains/categories/controllers/controller.py
@@ -7,7 +7,7 @@ from ftmcloud.api.domains.product_types.services.product_type_service import Pro
 
 from ftmcloud.models.patchdocument import PatchDocument
 from ftmcloud.models.response import Response, ResponseWithHttpInfo
-from ftmcloud.models.category import Category
+from ftmcloud.models.domains.categories.category import Category
 from ftmcloud.utils.views import controller
 
 categories_router = APIRouter()
