+++ ftmcloud/api/domains/categories/services/category_services.py
@@ -1,7 +1,7 @@
 from ftmcloud.core.exception.exception import FtmException
 from ftmcloud.core.service import Service
-from ftmcloud.models.category import Category
-from ftmcloud.models.product_type import ProductType
+from ftmcloud.models.domains.categories.category import Category
+from ftmcloud.models.domains.product_types.product_type import ProductType
 
 
 class CategoriesService(Service):
