+++ ftmcloud/api/domains/industries/services/industry_services.py
@@ -1,5 +1,5 @@
 from ftmcloud.core.service import Service
-from ftmcloud.models.industry import Industry
+from ftmcloud.models.domains.industries.industry import Industry
 
 
 class IndustriesService(Service):
