+++ ftmcloud/api/domains/privileges/services/privilege_services.py
@@ -1,5 +1,5 @@
 from ftmcloud.core.service import Service
-from ftmcloud.models.privilege import Privilege
+from ftmcloud.models.domains.privileges.privilege import Privilege
 
 
 class PrivilegesService(Service):
