+++ ftmcloud/api/domains/product_types/controllers/controller.py
@@ -6,7 +6,7 @@ from ftmcloud.api.domains.product_types.services.product_type_service import Pro
 
 from ftmcloud.models.patchdocument import PatchDocument
 from ftmcloud.models.response import Response, ResponseWithHttpInfo
-from ftmcloud.models.product_type import ProductType
+from ftmcloud.models.domains.product_types.product_type import ProductType
 from ftmcloud.utils.views import controller
 
 product_type_router = APIRouter()
