+++ ftmcloud/models/domains/industries/industry.py
