+++ ftmcloud/models/domains/product_types/product_type.py
@@ -1,6 +1,6 @@
 from pydantic.class_validators import Optional
 
-from ftmcloud.models.attribute import AttributeValue
+from ftmcloud.models.domains.attributes.attribute import AttributeValue
 from ftmcloud.models.document import BaseDocument
