--- ftmcloud/models/student.py
@@ -1,45 +1,0 @@
-from typing import Optional, Any
-
-from beanie import Document
-from pydantic import BaseModel, EmailStr
-
-
-class Student(Document):
-    fullname: str
-    email: EmailStr
-    course_of_study: str
-    year: int
-    gpa: float
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "fullname": "Abdulazeez Abdulazeez Adeshina",
-                "email": "abdul@school.com",
-                "course_of_study": "Water resources engineering",
-                "year": 4,
-                "gpa": "3.76"
-            }
-        }
-
-
-class UpdateStudentModel(BaseModel):
-    fullname: Optional[str]
-    email: Optional[EmailStr]
-    course_of_study: Optional[str]
-    year: Optional[int]
-    gpa: Optional[float]
-
-    class Settings:
-        name = "student"
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "fullname": "Abdulazeez Abdulazeez",
-                "email": "abdul@school.com",
-                "course_of_study": "Water resources and environmental engineering",
-                "year": 4,
-                "gpa": "5.0"
-            }
-        }
